# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{DEV} Pedia",
        "title": "{DEV} Pedia API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:9001",
    "basePath": "/",
    "paths": {
...
        "/api/articles/{source_type}/{id}/cluster": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get the cluster assignment for a specific article",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clustering"
                ],
                "summary": "Get article's cluster assignment",
                "parameters": [
                    {
                        "enum": [
                            "doc",
                            "git"
                        ],
                        "type": "string",
                        "description": "Source type",
                        "name": "source_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "label_propagation",
                        "description": "Clustering algorithm",
                        "name": "algorithm",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleClusterAssignment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/graph/clusters": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all detected clusters/communities in the knowledge graph",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clustering"
                ],
                "summary": "Get graph clusters",
                "parameters": [
                    {
                        "type": "string",
                        "default": "label_propagation",
                        "description": "Clustering algorithm",
                        "name": "algorithm",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.GetClustersResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/graph/clusters/run": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Detect communities in the knowledge graph using specified algorithm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clustering"
                ],
                "summary": "Run clustering algorithm",
                "parameters": [
                    {
                        "description": "Clustering request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.RunClusteringRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.RunClusteringResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fceek_dev-pedia_backend_internal_models.ArticleClusterAssignment": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "calculated_at": {
                    "type": "string"
                },
                "centrality_score": {
                    "type": "number"
                },
                "cluster_id": {
                    "type": "integer"
                },
                "cluster_label": {
                    "type": "string"
                }
            }
        },
...
        "fceek_dev-pedia_backend_internal_models.ArticleSourceType": {
            "type": "string",
            "enum": [
                "doc",
                "git"
            ],
            "x-enum-varnames": [
                "ArticleSourceDoc",
                "ArticleSourceGit"
            ]
        },
...
        "fceek_dev-pedia_backend_internal_models.ClusterInfo": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "avg_centrality": {
                    "type": "number"
                },
                "cluster_id": {
                    "type": "integer"
                },
                "density": {
                    "type": "number"
                },
                "label": {
                    "type": "string"
                },
                "representative_classification": {
                    "type": "integer"
                },
                "representative_id": {
                    "type": "string"
                },
                "representative_path": {
                    "type": "string"
                },
                "representative_source_type": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleSourceType"
                },
                "representative_title": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "fceek_dev-pedia_backend_internal_models.GetClustersResponse": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "clusters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ClusterInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "fceek_dev-pedia_backend_internal_models.RunClusteringRequest": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "description": "'label_propagation', etc.",
                    "type": "string"
                }
            }
        },
        "fceek_dev-pedia_backend_internal_models.RunClusteringResponse": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "cluster_count": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer token for regular authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GodToken": {
            "description": "God token for bootstrap operations",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}